Remove search filters from model memory
In the trades menu view a user can search for an item.
This search can also contain a filter.
Instead of saving the most recent search's filter in the model it is now passed
as a parameter. Especially since the search is saved with its filter
inside of the search history.